first pass at private for utils -> _internal
