+++ libs/langgraph/bench/sequential.py
@@ -1,7 +1,7 @@
 """Create a sequential no-op graph consisting of a few hundred nodes."""
 
+from langgraph._internal._runnable import RunnableCallable
 from langgraph.graph import MessagesState, StateGraph
-from langgraph.utils.runnable import RunnableCallable
 
 
 def create_sequential(number_nodes: int) -> StateGraph:
