+++ libs/langgraph/langgraph/pregel/algo.py
@@ -25,6 +25,7 @@ from langchain_core.callbacks.manager import AsyncParentRunManager, ParentRunMan
 from langchain_core.runnables.config import RunnableConfig
 from xxhash import xxh3_128_hexdigest
 
+from langgraph._internal._config import merge_configs, patch_config
 from langgraph.channels.base import BaseChannel
 from langgraph.channels.topic import Topic
 from langgraph.checkpoint.base import (
@@ -64,7 +65,6 @@ from langgraph.constants import (
     RETURN,
     TAG_HIDDEN,
     TASKS,
-    Send,
 )
 from langgraph.managed.base import ManagedValueMapping
 from langgraph.pregel.call import get_runnable_for_task, identifier
@@ -80,8 +80,8 @@ from langgraph.types import (
     PregelScratchpad,
     PregelTask,
     RetryPolicy,
+    Send,
 )
-from langgraph.utils.config import merge_configs, patch_config
 
 GetNextVersion = Callable[[Optional[V], None], V]
 SUPPORTS_EXC_NOTES = sys.version_info >= (3, 11)
