+++ libs/langgraph/langgraph/pregel/read.py
@@ -10,13 +10,13 @@ from typing import (
 
 from langchain_core.runnables import Runnable, RunnableConfig
 
+from langgraph._internal._config import merge_configs
+from langgraph._internal._runnable import RunnableCallable, RunnableSeq
 from langgraph.constants import CONF, CONFIG_KEY_READ
 from langgraph.pregel.protocol import PregelProtocol
 from langgraph.pregel.utils import find_subgraph_pregel
 from langgraph.pregel.write import ChannelWrite
 from langgraph.types import CachePolicy, RetryPolicy
-from langgraph.utils.config import merge_configs
-from langgraph.utils.runnable import RunnableCallable, RunnableSeq
 
 READ_TYPE = Callable[[Union[str, Sequence[str]], bool], Union[Any, dict[str, Any]]]
 INPUT_CACHE_KEY_TYPE = tuple[Callable[..., Any], tuple[str, ...]]
